moved commandline into a module #495
